HealthM3/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
pymongo>=4.0.0
python-dotenv>=0.19.0
requests>=2.26.0
flask-cors>=3.0.10
gunicorn>=20.1.0
pymysql>=1.0.2
scipy>=1.10.0
python-dateutil>=2.8.2
tqdm>=4.65.0
jieba>=0.42.1
snownlp>=0.12.3 